requirements.txt
# Clickstream Analytics Pipeline - Python Dependencies
# Install with: pip install -r requirements.txt
# Python 3.9 compatible versions

# Core Data Processing
pyspark==3.5.1
great-expectations==0.18.21
delta-spark==3.1.0
pyarrow==15.0.0

# Data Generation
faker==23.1.0
tqdm==4.66.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# AWS/S3/MinIO
boto3==1.34.34
s3fs==2023.12.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
click==8.1.7

# Testing
pytest==8.0.0
pytest-cov==4.1.0